Split h.resources into h.traversal
Change the `h/resources.py` module into an `h/traversal` package that is
internally organised into separate `roots.py` and `contexts.py` modules.

This is part of the refactoring described in this post:

https://hypothes-is.slack.com/files/U074DEU66/FA511033L/h_traversal__resources__refactoring

The name `h.traversal` is preferred over `h.resources` because:

* "Traversal" is the high-level name of the Pyramid feature being used
  here (the relevant Pyramid documentation chapter titles refer to
  "traversal", for example)

* "Resources" is a very generic and overloaded term, not clear what it
  refers to

* The classes inside `h.resources` (now `h.traversal`) aren't called
  "resource" classes in our code, we call them all `*Root` or `*Context`,
  and don't use the term "resource"

The single module has been split into separate roots and contexts
modules because:

* It keeps the modules and their associated test modules smaller

* roots.py and contexts.py provide convenient places to put module-level
  docstrings explaining what roots and contexts are and how they're used
  in h
