+++ tests/h/routes_test.py
@@ -34,7 +34,7 @@ def test_includeme():
         call('claim_account_legacy', '/claim_account/{token}'),
         call('dismiss_sidebar_tutorial', '/app/dismiss_sidebar_tutorial'),
         call('activity.search', '/search'),
-        call('activity.user_search', '/users/{username}', factory='h.resources:UserRoot', traverse='/{username}'),
+        call('activity.user_search', '/users/{username}', factory='h.traversal:UserRoot', traverse='/{username}'),
         call('admin_index', '/admin/'),
         call('admin_admins', '/admin/admins'),
         call('admin_badge', '/admin/badge'),
@@ -50,23 +50,23 @@ def test_includeme():
         call('admin_nipsa', '/admin/nipsa'),
         call('admin_oauthclients', '/admin/oauthclients'),
         call('admin_oauthclients_create', '/admin/oauthclients/new'),
-        call('admin_oauthclients_edit', '/admin/oauthclients/{id}', factory='h.resources.AuthClientRoot', traverse='/{id}'),
+        call('admin_oauthclients_edit', '/admin/oauthclients/{id}', factory='h.traversal.AuthClientRoot', traverse='/{id}'),
         call('admin_organizations', '/admin/organizations'),
         call('admin_organizations_create', '/admin/organizations/new'),
         call('admin_organizations_delete',
              '/admin/organizations/delete/{pubid}',
-             factory='h.resources.OrganizationRoot',
+             factory='h.traversal.OrganizationRoot',
              traverse='/{pubid}'),
         call('admin_organizations_edit',
              '/admin/organizations/{pubid}',
-             factory='h.resources.OrganizationRoot',
+             factory='h.traversal.OrganizationRoot',
              traverse='/{pubid}'),
         call('admin_staff', '/admin/staff'),
         call('admin_users', '/admin/users'),
         call('admin_users_activate', '/admin/users/activate'),
         call('admin_users_delete', '/admin/users/delete'),
         call('admin_users_rename', '/admin/users/rename'),
-        call('annotation', '/a/{id}', factory='h.resources:AnnotationRoot', traverse='/{id}'),
+        call('annotation', '/a/{id}', factory='h.traversal:AnnotationRoot', traverse='/{id}'),
         call('stream', '/stream'),
         call('stream.user_query', '/u/{user}'),
         call('stream.tag_query', '/t/{tag}'),
@@ -76,23 +76,23 @@ def test_includeme():
         call('api.annotations', '/api/annotations'),
         call('api.annotation',
              '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}',
-             factory='h.resources:AnnotationRoot',
+             factory='h.traversal:AnnotationRoot',
              traverse='/{id}'),
         call('api.annotation_flag',
              '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}/flag',
-             factory='h.resources:AnnotationRoot',
+             factory='h.traversal:AnnotationRoot',
              traverse='/{id}'),
         call('api.annotation_hide',
              '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}/hide',
-             factory='h.resources:AnnotationRoot',
+             factory='h.traversal:AnnotationRoot',
              traverse='/{id}'),
         call('api.annotation.jsonld',
              '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}.jsonld',
-             factory='h.resources:AnnotationRoot',
+             factory='h.traversal:AnnotationRoot',
              traverse='/{id}'),
         call('api.profile', '/api/profile'),
         call('api.debug_token', '/api/debug-token'),
-        call('api.group_member', '/api/groups/{pubid}/members/{user}', factory='h.resources.GroupRoot', traverse='/{pubid}'),
+        call('api.group_member', '/api/groups/{pubid}/members/{user}', factory='h.traversal.GroupRoot', traverse='/{pubid}'),
         call('api.search', '/api/search'),
         call('api.users', '/api/users'),
         call('api.user', '/api/users/{username}'),
@@ -105,12 +105,12 @@ def test_includeme():
         call('stream_atom', '/stream.atom'),
         call('stream_rss', '/stream.rss'),
         call('organization_logo', '/organizations/{pubid}/logo',
-             factory='h.resources.OrganizationLogoRoot', traverse='/{pubid}'),
+             factory='h.traversal.OrganizationLogoRoot', traverse='/{pubid}'),
         call('api.groups', '/api/groups'),
         call('group_create', '/groups/new'),
-        call('group_edit', '/groups/{pubid}/edit', factory='h.resources.GroupRoot', traverse='/{pubid}'),
-        call('group_read', '/groups/{pubid}/{slug:[^/]*}', factory='h.resources.GroupRoot', traverse='/{pubid}'),
-        call('group_read_noslug', '/groups/{pubid}', factory='h.resources.GroupRoot', traverse='/{pubid}'),
+        call('group_edit', '/groups/{pubid}/edit', factory='h.traversal.GroupRoot', traverse='/{pubid}'),
+        call('group_read', '/groups/{pubid}/{slug:[^/]*}', factory='h.traversal.GroupRoot', traverse='/{pubid}'),
+        call('group_read_noslug', '/groups/{pubid}', factory='h.traversal.GroupRoot', traverse='/{pubid}'),
         call('help', '/docs/help'),
         call('onboarding', '/welcome/'),
         call('custom_onboarding', '/welcome/{slug}'),
